Add new and edit form components, add api response checkers on new and edit forms, add some interactive after acts
